+++ app/src/main/java/com/example/hrm/db/entity/Ecg.kt
@@ -18,9 +18,9 @@ import androidx.room.PrimaryKey
     indices = [Index(value = ["sessionId"])]
 )
 data class Ecg(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
 
-    val sessionId: Int? = null,        // 外键，关联体检记录（可选）
+    val sessionId: Long? = null,        // 外键，关联体检记录（可选）
     val date: Long = System.currentTimeMillis(), // 检查时间
 
     val imagePath: String?,            // 心电图图像文件路径（本地路径或 URI）
